Move handling for Counting settings to own file
